% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AORC_meterodf}
\alias{AORC_meterodf}
\title{AORC Variable Metadata}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{ClimateVar}{Human-readable climate variable name.}
\item{VarLabel}{Variable name used in AORC data products.}
\item{Units}{Measurement units of the variable.}
}
}
\source{
NOAA Analysis of Record for Calibration (AORC)
}
\usage{
AORC_meterodf
}
\description{
Lookup table describing meteorological variables available in the
Analysis of Record for Calibration (AORC) forcing dataset, including
variable labels used in AORC files and their associated units.
}
\details{
This dataset can be used to identify required climate variables when
preparing AORC data for use in microclimate modeling workflows.

Variables included are:
\itemize{
\item Total precipitation
\item Air temperature
\item Specific humidity
\item Downward longwave radiation
\item Downward shortwave radiation
\item Atmospheric pressure
\item East-west wind component (U)
\item South-north wind component (V)
\item Diffuse radiation
}
}
\keyword{datasets}
