% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmask_bits}
\alias{fmask_bits}
\title{HLS Fmask Bit Description Table}
\format{
A data frame with 16 rows and 4 columns:
\describe{
\item{Mask_name}{Name of the mask category}
\item{Bit_Position}{Bit position in the Fmask band}
\item{BitValue}{Bit value corresponding to the description}
\item{description}{Human readable description of the bit value}
}
}
\source{
\url{https://lpdaac.usgs.gov/documents/1698/HLS_User_Guide_V2.pdf}
}
\usage{
fmask_bits
}
\description{
A reference data frame describing the bit positions and values used in the
Fmask quality band of the NASA Harmonized Landsat Sentinel-2 (HLS) V2.0
product. Used internally by download_hls() for cloud and quality masking.
}
\keyword{datasets}
