% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FparLAI_QC}
\alias{FparLAI_QC}
\title{MODIS LAI/FPAR QC Bit Description Table}
\format{
A data frame with 15 rows and 4 columns:
\describe{
\item{Mask_name}{Name of the mask category}
\item{Bit_Position}{Bit position in the FparLai_QC band}
\item{BitValue}{Bit value corresponding to the description}
\item{description}{Human readable description of the bit value}
}
}
\source{
\url{https://lpdaac.usgs.gov/documents/624/MOD15_User_Guide_V6.pdf}
}
\usage{
FparLAI_QC
}
\description{
A reference data frame describing the bit positions and values used in the
FparLai_QC quality band of the MODIS MCD15A3H LAI/FPAR product. Used
internally by download_modis_lai() for quality masking.
}
\keyword{datasets}
