% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microclim_meterodf}
\alias{Microclim_meterodf}
\title{Microclim Meteorological Variable Metadata}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{ClimateVar}{Human-readable climate variable name.}
\item{VarLabel}{Standardized variable name used by Microclim.}
\item{Units}{Measurement units of the variable.}
}
}
\usage{
Microclim_meterodf
}
\description{
Lookup table describing meteorological variables and naming conventions
used by the microclimate modeling functions in this package.
}
\details{
This dataset provides standardized variable names and units expected by
Microclim-compatible forcing datasets.

Variables included are:
\itemize{
\item Total precipitation
\item Air temperature
\item Relative humidity
\item Downward longwave radiation
\item Downward shortwave radiation
\item Atmospheric pressure
\item Wind speed
\item Wind direction
\item Diffuse radiation
}
}
\seealso{
\code{\link{AORC_meterodf}}
}
\keyword{datasets}
